[#4] chore: DishCell의 badgeView 보여지는 방식 수정 및 DiscountType 분리
